hook up dpid fix to publish flow
